+++ python/shipDet_conf.py
@@ -155,9 +155,9 @@ def configure_snd_siliconTarget(yaml_file, ship_geo):
     # Initialize detector
     if ship_geo.SiliconTarget_geo.zPosition == "auto":
         # Get the the center of the next to last magnet (temporary placement)
-        # Offset placement of detector by 130 cm, magnet is 2* 212.54 cm,
-        # 120 layers at 132 cm will fit.
-        ship_geo.SiliconTarget_geo.zPosition = find_shield_center(ship_geo)[2][-2] + 130
+        # Offset placement of detector by 140 cm, magnet is 2* 212.54 cm,
+        # 120 layers at 132 cm will fit, with 140 cm offset final layer within 10 cm of MTC.
+        ship_geo.SiliconTarget_geo.zPosition = find_shield_center(ship_geo)[2][-2] + 140
         print("SiliconTarget zPosition set to ", ship_geo.SiliconTarget_geo.zPosition)
     SiliconTarget = ROOT.SiliconTarget("SiliconTarget", ROOT.kTRUE)
     SiliconTarget.SetSiliconTargetParameters(